+++ src/bbs2gh/Program.cs
@@ -112,6 +112,12 @@ namespace OctoshiftCLI.BbsToGithub
 
         private static async Task LatestVersionCheck(ServiceProvider sp)
         {
+            if (Environment.GetEnvironmentVariable("GEI_SKIP_VERSION_CHECK")?.ToUpperInvariant() == "TRUE")
+            {
+                Logger.LogInformation("Skipped latest version check of the bbs2gh extension");
+                return;
+            }
+
             var versionChecker = sp.GetRequiredService<VersionChecker>();
 
             if (await versionChecker.IsLatest())