+++ RELEASENOTES.md
@@ -2,3 +2,4 @@
 - Allow CLI to fail fast when an unauthorized token is provided by preventing retry logic on 401 errors
 - Improve performance when reclaiming a single mannequin with `reclaim-mannequin`
 - Improve logging for `reclaim-mannequin` command
+- Skip version checking on startup if environment variable `GEI_SKIP_VERSION_CHECK` is set to true.